# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.",
        "title": "eID based AI enabled Government Template Platform V3.0 API",
        "termsOfService": "https://github.com/snykk/go-rest-boilerplate",
        "contact": {
            "name": "eID based AI enabled Government Template Platform V3.0",
            "url": "https://github.com/snykk/go-rest-boilerplate"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/integrations/{provider}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Тухайн хэрэглэгчийн нэг провайдерын хадгалсан токеныг устгана (idempotent).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "Интеграцийг салгах",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider id (google-drive, dropbox, google-meet)",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disconnected",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Unknown provider",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/sso/native": {
            "post": {
                "description": "Mobile app (iOS/Android) ASWebAuthenticationSession-ийн PKCE code-ийг public client-ээр (client_secret-гүй, code_verifier-тэй) солин, иргэнийг upsert хийж JWT хос олгоно. State шалгалтгүй (PKCE хамгаална). BFF нь token/refresh_token-ийг httpOnly cookie-д суулгана.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sso"
                ],
                "summary": "Gerege SSO native (mobile PKCE) нэвтрэлт",
                "parameters": [
                    {
                        "description": "Native PKCE code exchange",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.SSONativeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Missing code/code_verifier",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "template_internal_http_datatransfers_requests.SSONativeRequest": {
            "type": "object",
            "required": [
                "code",
                "code_verifier",
                "redirect_uri"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "code_verifier": {
                    "type": "string"
                },
                "redirect_uri": {
                    "description": "RedirectURI нь native client-д бүртгэгдсэн (жишээ\ngeregetemp://oauth2/callback) байх ёстой — code exchange-д яг тааруулна.",
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_handlers_v1.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "/auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен (access token).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}